legacy_version/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-multipart
google-genai
groq
mistralai
cohere
openai
python-dotenv
aiofiles
